Extract music player logic to its own class
